requirements.txt
numpy
pandas
pinecone-client>=3.0.0
pyarrow
qdrant_client
tqdm
python-dotenv
huggingface_hub
pymilvus>=2.3.6
openai>=1.10.0
litellm>=1.20.0
google-cloud-aiplatform
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
ratelimit
backoff
kdbai-client
sentry-sdk[opentelemetry]
halo
sentence-transformers>=2.6.1
pyvespa
lancedb
chromadb
txtai
weaviate-client>=4.0.0
astrapy
cassandra-driver
tenacity
torch
twine
datasets~=2.16,>=2.19.0
mlx_embedding_models
azure-search-documents
azure-identity